Toolchain update
Swapped prepros out for an Atom plugin, as prepros was getting a bit annoying with all those popups. `node-sass` is still used, but some parameters were changed. For all intents and purposes, this commit should most probably be ignored as nothing substantial was changed. The only thing that is a bit different is that the encoding at the start of both .min.css files is now removed. If this somehow ends up causing trouble, then DO refer to this commit in your issue, if you read this!