Updated Nord flavor
The Nord flavor has been updated to the latest version. Alerts and the .nano progress bar have been removed, the .toast has been added using the color palette that the flavor uses (as far as I understand it), the alert colors have been move to .card color variants, .drawer has been added using the colors of <nav>, generic shadows remain unchanged, based on the fact that the flavor still uses shadows, the .drawer component does not have a shadow by the way (I think it causes some problems, but this requires further testing). If anyone wants to update the flavor, please refer to this commit!